Убрал Worker.java. Все перенес в ServerHandler.java
